COMMERCE-11740 autogenerated code (gw buildService)

+++ modules/apps/commerce/commerce-inventory-api/src/main/java/com/liferay/commerce/inventory/model/CommerceInventoryAuditTable.java
@@ -8,6 +8,8 @@ package com.liferay.commerce.inventory.model;
 import com.liferay.petra.sql.dsl.Column;
 import com.liferay.petra.sql.dsl.base.BaseTable;
 
+import java.math.BigDecimal;
+
 import java.sql.Clob;
 import java.sql.Types;
 
@@ -52,9 +54,9 @@ public class CommerceInventoryAuditTable
 	public final Column<CommerceInventoryAuditTable, Clob> logTypeSettings =
 		createColumn(
 			"logTypeSettings", Clob.class, Types.CLOB, Column.FLAG_DEFAULT);
-	public final Column<CommerceInventoryAuditTable, Integer> quantity =
+	public final Column<CommerceInventoryAuditTable, BigDecimal> quantity =
 		createColumn(
-			"quantity", Integer.class, Types.INTEGER, Column.FLAG_DEFAULT);
+			"quantity", BigDecimal.class, Types.DECIMAL, Column.FLAG_DEFAULT);
 	public final Column<CommerceInventoryAuditTable, String> sku = createColumn(
 		"sku", String.class, Types.VARCHAR, Column.FLAG_DEFAULT);
 	public final Column<CommerceInventoryAuditTable, String> unitOfMeasureKey =
